requirements.txt
aiogram>=3.0
pydub>=0.25
python-dotenv>=1.0
audioop-lts>=0.2
aiohttp>=3.9
aiohttp-session[secure]>=2.12
cryptography>=42.0
